modify data_reps datatype

+++ libs/core/iostream/tests/unit/seekable_filter.cpp
@@ -12,8 +12,8 @@
 #include <hpx/hpx_main.hpp>
 #include <hpx/modules/iostream.hpp>
 #include <hpx/modules/testing.hpp>
-#include <cstddef>
 
+#include <cstddef>
 #include <iosfwd>
 #include <vector>
 
@@ -29,8 +29,7 @@ using namespace hpx::iostream::test;
 void seekable_filter_test()
 {
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '0');
+        vector<char> test(data_reps * data_length(), '0');
         filtering_stream<seekable> io;
         io.push(identity_seekable_filter());
         io.push(container_device<vector<char>>(test));
@@ -40,8 +39,7 @@ void seekable_filter_test()
     }
 
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '0');
+        vector<char> test(data_reps * data_length(), '0');
         filtering_stream<seekable> io;
         io.push(identity_seekable_filter());
         io.push(container_device<vector<char>>(test));
@@ -51,8 +49,7 @@ void seekable_filter_test()
     }
 
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '0');
+        vector<char> test(data_reps * data_length(), '0');
         filtering_stream<seekable> io;
         io.push(identity_seekable_multichar_filter());
         io.push(container_device<vector<char>>(test));
@@ -62,8 +59,7 @@ void seekable_filter_test()
     }
 
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '0');
+        vector<char> test(data_reps * data_length(), '0');
         filtering_stream<seekable> io;
         io.push(identity_seekable_multichar_filter());
         io.push(container_device<vector<char>>(test));
